requirements.txt
ipykernel
python-dotenv
openai
re 
json